LayoutTests/imported/w3c/web-platform-tests/css/css-contain/parsing/contain-computed-expected.txt


PASS Property contain value 'none'
PASS Property contain value 'strict'
PASS Property contain value 'content'
PASS Property contain value 'size'
PASS Property contain value 'layout'
PASS Property contain value 'style'
PASS Property contain value 'paint'
PASS Property contain value 'size layout'
PASS Property contain value 'style paint'
FAIL Property contain value 'style layout paint' assert_equals: expected "content" but got "layout style paint"
FAIL Property contain value 'size style layout paint' assert_equals: expected "strict" but got "size layout style paint"
FAIL Property contain value 'size layout paint' assert_equals: expected "size layout paint" but got "strict"
FAIL Property contain value 'layout paint' assert_equals: expected "layout paint" but got "content"
